full reorg, sample runs
