+++ activitysim/abm/models/__init__.py
@@ -13,6 +13,8 @@ from . import cdap
 from . import free_parking
 from . import generate_beam_plans
 from . import initialize
+from . import initialize_from_usim
+from . import initialize_skims_from_beam
 from . import joint_tour_composition
 from . import joint_tour_destination
 from . import joint_tour_frequency
