+++ austin_mp/configs/settings.yaml
@@ -1,6 +1,17 @@
 
 inherit_settings: True
 
+beam_skims_url: https://beam-outputs.s3.amazonaws.com/output/austin/austin-prod-200k-skims-with-h3-index-final__2020-04-18_09-44-24_wga/ITERS/it.0/0.skimsOD.UrbanSim.Full.csv.gz
+
+create_skims_from_beam: True
+usim_zone_geoms: h3
+# usim_zone_shapefile:
+
+create_inputs_from_usim_data: True
+usim_data_store: model_data.h5
+
+
+
 # raise error if any sub-process fails without waiting for others to complete
 fail_fast: True
 
@@ -51,7 +62,7 @@ trace_od:
 #trace_od: [5, 11]
 
 # to resume after last successful checkpoint, specify resume_after: _
-resume_after: trip_mode_choice
+# resume_after: trip_mode_choice
 
 models:
   ### mp_initialize step
@@ -116,5 +127,3 @@ output_tables:
     - tours
     - trips
     - beam_plans
-
-
