+++ austin_mp/simulation_mp.py
@@ -5,7 +5,6 @@ from __future__ import (absolute_import, division, print_function, )
 from future.standard_library import install_aliases
 install_aliases()  # noqa: E402
 
-import sys
 import logging
 
 from activitysim.core import mem
@@ -15,16 +14,15 @@ from activitysim.core import config
 from activitysim.core import pipeline
 from activitysim.core import mp_tasks
 from activitysim.core import chunk
-from os import path
 
 
 # Create tables if they do not exist already
-persons_table = path.exists("data/persons.csv")
-households_table = path.exists("data/households.csv")
-land_use_table = path.exists("data/land_use.csv")
+# persons_table = path.exists("data/persons.csv")
+# households_table = path.exists("data/households.csv")
+# land_use_table = path.exists("data/land_use.csv")
 
-if not persons_table & households_table & land_use_table:
-    from data import usim_tables
+# if not persons_table & households_table & land_use_table:
+#     from data import usim_tables
 
 
 logger = logging.getLogger('activitysim')
@@ -45,6 +43,18 @@ def cleanup_output_files():
 
 
 def run(run_list, injectables=None):
+
+    # Create a new skims.omx file from BEAM (http://beam.lbl.gov/) skims
+    # if skims do not already exist in the input data directory
+    if config.setting('create_skims_from_beam'):
+        pipeline.run(models=['create_skims_from_beam'])
+        pipeline.close_pipeline()
+
+    # Create persons, households, and land use .csv files from UrbanSim
+    # data if these files do not already exist in the input data directory
+    if config.setting('create_inputs_from_usim_data'):
+        pipeline.run(models=['load_usim_data', 'create_inputs_from_usim_data'])
+        pipeline.close_pipeline()
 
     if run_list['multiprocess']:
         logger.info("run multiprocess simulation")
